+++ scripts/vars.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# Add In Vars From .env File
+export $(cat .env | sed 's/#.*//g' | xargs)